Completed day 23 p2
Fixed defect in min_manhattan_dist to origin as well as determined that eligible sub bounds could in fact contain all nanos, but a subset of their bounds would not. The search algorithm now searches the min manhattan dist sub bounds and if it has no eligible subbounds proceeds to the next.